+++ de.metas.adempiere.adempiere/base/src/main/java-gen/org/compiere/model/X_C_BPartner_Product.java
@@ -1,19 +1,3 @@
-/******************************************************************************
- * Product: Adempiere ERP & CRM Smart Business Solution                       *
- * Copyright (C) 1999-2007 ComPiere, Inc. All Rights Reserved.                *
- * This program is free software, you can redistribute it and/or modify it    *
- * under the terms version 2 of the GNU General Public License as published   *
- * by the Free Software Foundation. This program is distributed in the hope   *
- * that it will be useful, but WITHOUT ANY WARRANTY, without even the implied *
- * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.           *
- * See the GNU General Public License for more details.                       *
- * You should have received a copy of the GNU General Public License along    *
- * with this program, if not, write to the Free Software Foundation, Inc.,    *
- * 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA.                     *
- * For the text or an alternative of this public license, you may reach us    *
- * ComPiere, Inc., 2620 Augustine Dr. #245, Santa Clara, CA 95054, USA        *
- * or via info@compiere.org or http://www.compiere.org/license.html           *
- *****************************************************************************/
 /** Generated Model - DO NOT CHANGE */
 package org.compiere.model;
 
@@ -32,7 +16,7 @@ public class X_C_BPartner_Product extends org.compiere.model.PO implements I_C_B
 	/**
 	 *
 	 */
-	private static final long serialVersionUID = -1333517629L;
+	private static final long serialVersionUID = 674564741L;
 
     /** Standard Constructor */
     public X_C_BPartner_Product (Properties ctx, int C_BPartner_Product_ID, String trxName)
@@ -41,7 +25,10 @@ public class X_C_BPartner_Product extends org.compiere.model.PO implements I_C_B
       /** if (C_BPartner_Product_ID == 0)
         {
 			setC_BPartner_ID (0);
+			setC_BPartner_Product_ID (0);
 			setIsCurrentVendor (false);
+// N
+			setIsDropShip (false);
 // N
 			setM_Product_ID (0);
 			setShelfLifeMinDays (0);
@@ -55,29 +42,13 @@ public class X_C_BPartner_Product extends org.compiere.model.PO implements I_C_B
       super (ctx, rs, trxName);
     }
 
-    /** AccessLevel
-      * @return 3 - Client - Org 
-      */
-    @Override
-    protected int get_AccessLevel()
-    {
-      return accessLevel.intValue();
-    }
 
     /** Load Meta Data */
     @Override
     protected org.compiere.model.POInfo initPO (Properties ctx)
     {
-      org.compiere.model.POInfo poi = org.compiere.model.POInfo.getPOInfo (ctx, Table_ID, get_TrxName());
+      org.compiere.model.POInfo poi = org.compiere.model.POInfo.getPOInfo (ctx, Table_Name, get_TrxName());
       return poi;
-    }
-
-    @Override
-    public String toString()
-    {
-      StringBuilder sb = new StringBuilder ("X_C_BPartner_Product[")
-        .append(get_ID()).append("]");
-      return sb.toString();
     }
 
 	@Override
@@ -112,6 +83,28 @@ public class X_C_BPartner_Product extends org.compiere.model.PO implements I_C_B
 	public int getC_BPartner_ID () 
 	{
 		Integer ii = (Integer)get_Value(COLUMNNAME_C_BPartner_ID);
+		if (ii == null)
+			 return 0;
+		return ii.intValue();
+	}
+
+	/** Set Geschäftspartner-Produkt.
+		@param C_BPartner_Product_ID Geschäftspartner-Produkt	  */
+	@Override
+	public void setC_BPartner_Product_ID (int C_BPartner_Product_ID)
+	{
+		if (C_BPartner_Product_ID < 1) 
+			set_ValueNoCheck (COLUMNNAME_C_BPartner_Product_ID, null);
+		else 
+			set_ValueNoCheck (COLUMNNAME_C_BPartner_Product_ID, Integer.valueOf(C_BPartner_Product_ID));
+	}
+
+	/** Get Geschäftspartner-Produkt.
+		@return Geschäftspartner-Produkt	  */
+	@Override
+	public int getC_BPartner_Product_ID () 
+	{
+		Integer ii = (Integer)get_Value(COLUMNNAME_C_BPartner_Product_ID);
 		if (ii == null)
 			 return 0;
 		return ii.intValue();
@@ -206,6 +199,32 @@ public class X_C_BPartner_Product extends org.compiere.model.PO implements I_C_B
 	public boolean isCurrentVendor () 
 	{
 		Object oo = get_Value(COLUMNNAME_IsCurrentVendor);
+		if (oo != null) 
+		{
+			 if (oo instanceof Boolean) 
+				 return ((Boolean)oo).booleanValue(); 
+			return "Y".equals(oo);
+		}
+		return false;
+	}
+
+	/** Set Streckengeschäft.
+		@param IsDropShip 
+		Beim Streckengeschäft wird die Ware direkt vom Lieferanten zum Kunden geliefert
+	  */
+	@Override
+	public void setIsDropShip (boolean IsDropShip)
+	{
+		set_Value (COLUMNNAME_IsDropShip, Boolean.valueOf(IsDropShip));
+	}
+
+	/** Get Streckengeschäft.
+		@return Beim Streckengeschäft wird die Ware direkt vom Lieferanten zum Kunden geliefert
+	  */
+	@Override
+	public boolean isDropShip () 
+	{
+		Object oo = get_Value(COLUMNNAME_IsDropShip);
 		if (oo != null) 
 		{
 			 if (oo instanceof Boolean)