+++ de.metas.adempiere.adempiere/base/src/main/java/org/adempiere/bpartner/service/IBPartnerStatisticsUpdater.java
@@ -13,15 +13,14 @@ package org.adempiere.bpartner.service;
  * 
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
  * 
  * You should have received a copy of the GNU General Public
- * License along with this program.  If not, see
+ * License along with this program. If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
  * #L%
  */
-
 
 import java.util.Properties;
 import java.util.Set;
@@ -29,18 +28,25 @@ import java.util.Set;
 import org.adempiere.ad.persistence.ModelDynAttributeAccessor;
 import org.adempiere.util.ISingletonService;
 import org.compiere.model.I_C_AllocationHdr;
-import org.compiere.model.I_C_BPartner;
+import org.compiere.model.I_C_BPartner_Stats;
 
 /**
- * Service used to update {@link I_C_BPartner#COLUMN_TotalOpenBalance} and {@link I_C_BPartner#COLUMN_SO_CreditUsed}.
+ * Service used to update {@link I_C_BPartner_Stats#COLUMN_TotalOpenBalance}, {@link I_C_BPartner_Stats#COLUMN_SO_CreditUsed}, {@link I_C_BPartner_Stats#COLUMN_ActualLifeTimeValue} and {@link I_C_BPartner_Stats#COLUMN_SOCreditStatus}
  * 
  * @author tsa
  *
  */
-public interface IBPartnerTotalOpenBalanceUpdater extends ISingletonService
+public interface IBPartnerStatisticsUpdater extends ISingletonService
 {
 	ModelDynAttributeAccessor<I_C_AllocationHdr, Boolean> DYNATTR_DisableUpdateTotalOpenBalances = new ModelDynAttributeAccessor<>("org.adempiere.bpartner.service.IBPartnerTotalOpenBalanceUpdater.DisableUpdateTotalOpenBalances", Boolean.class);
-	
-	void updateTotalOpenBalances(final Properties ctx, final Set<Integer> bpartnerToBalances, final String trxName);
+
+	/**
+	 * This method will update all the bpartner statistics based on the legacy sql.
+	 * 
+	 * @param ctx
+	 * @param bpartnerIDs
+	 * @param trxName
+	 */
+	void updateBPartnerStatistics(final Properties ctx, final Set<Integer> bpartnerIDs, final String trxName);
 
 }