+++ de.metas.adempiere.adempiere/migration/src/main/sql/postgresql/system/10-de.metas.adempiere/5446150_sys_FRESH-254_set_defaultvalue_to_isManuall_in_returnables.sql
@@ -1,0 +1,10 @@
+-- 25.05.2016 15:27
+-- I forgot to set the DICTIONARY_ID_COMMENTS System Configurator
+UPDATE AD_Field SET DefaultValue='Y',Updated=TO_TIMESTAMP('2016-05-25 15:27:10','YYYY-MM-DD HH24:MI:SS'),UpdatedBy=100 WHERE AD_Field_ID=555572
+;
+
+-- 25.05.2016 15:33
+-- I forgot to set the DICTIONARY_ID_COMMENTS System Configurator
+UPDATE AD_Field SET DefaultValue='Y',Updated=TO_TIMESTAMP('2016-05-25 15:33:59','YYYY-MM-DD HH24:MI:SS'),UpdatedBy=100 WHERE AD_Field_ID=555573
+;
+